Reset modified state on form submit
Submiting did not make this.state.modified resetting to false again.

Issue -> For example, when passing Prompt component (react-router-dom) like so

`<Formol extra={(this.state) => (<Prompt when={this.state.modified} />)}`

It will prompt the user a warning message to confirm the redirection after form submitting. This hangs
the redirection waiting for the user confirmation.

but the confirmation is implicit as soon as he clicks the submit button. Therefore there is no need to let the form in a modified state when
submiting was successful.